feat: Add LRUCache implementation
This commit adds a new file `3-lru_cache.py` which implements the Least Recently Used (LRU) caching algorithm. It extends the `BaseCaching` class and provides the `put` and `get` methods for caching and retrieving items respectively. The `put` method discards the least recently used item when the cache is full.