feat: Add MRUCache implementation
This commit adds a new file `4-mru_cache.py` which implements the Most Recently Used (MRU) caching algorithm. It extends the `BaseCaching` class and provides the `put` and `get` methods for caching and retrieving items respectively. The `put` method discards the least recently used item when the cache is full.