feat: Add LFUCache implementation
This commit adds a new file `100-lfu_cache.py` which implements the Least Frequently Used (LFU) caching algorithm. It extends the `BaseCaching` class and provides the `put` and `get` methods for caching and retrieving items respectively. The `put` method discards the least frequently used item when the cache is full.